Refactoring
* Extract creation of shutdown signal listener
