+++ src/helpers/setProcessShutdownSignalListeners.js
@@ -1,0 +1,13 @@
+/* eslint-disable no-process-exit,no-console */
+const createShutdownSignalListener = require('./createShutdownSignalListener')
+
+module.exports = (
+  shutdown,
+  logger = console,
+  forceExitTime = 20,
+) => {
+  const shutdownListener = createShutdownSignalListener(shutdown, logger, forceExitTime)
+
+  process.on('SIGTERM', shutdownListener)
+  process.on('SIGINT', shutdownListener)
+}
